# docs/openapi/saas-platform.yaml
openapi: 3.1.0
info:
  title: BettaFish SaaS Platform API
  version: 0.1.0
  summary: Contract for the BettaFish SaaS service layer.
  description: |
    This contract is the target API for the Next.js SaaS console. It intentionally
    does not mirror the legacy Flask + Streamlit routes one-for-one. Backend
    implementations may adapt existing endpoints such as `/api/status`,
    `/api/config`, `/api/search`, and `/api/report/*`, but the frontend should
    call the versioned SaaS resources below.
servers:
  - url: http://localhost:8000/api/v1
    description: Local backend service
  - url: https://api.bettafish.example.com/api/v1
    description: SaaS deployment placeholder
security:
  - bearerAuth: []
...
paths:
...
  /system/components:
    get:
      tags: [System]
      summary: List engine component runtime states.
      operationId: listSystemComponents
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Component states.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemComponentsResponse"
  /system/components/{componentId}:start:
    post:
      tags: [System]
      summary: Start a component.
      operationId: startSystemComponent
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ComponentId"
      responses:
        "202":
          description: Start requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentActionResponse"
        default:
          $ref: "#/components/responses/Error"
  /system/components/{componentId}:stop:
    post:
      tags: [System]
      summary: Stop a component.
      operationId: stopSystemComponent
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ComponentId"
      responses:
        "202":
          description: Stop requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentActionResponse"
        default:
          $ref: "#/components/responses/Error"
  /system/config:
    get:
      tags: [System]
      summary: Read system configuration with sensitive values masked.
      operationId: getSystemConfig
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: includeSchema
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Masked config values and editable field metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigResponse"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [System]
      summary: Update system configuration.
      description: |
        Sensitive fields may be omitted to keep current values. A masked value
        such as `********` must be ignored by the backend and never persisted as
        the real secret.
      operationId: updateSystemConfig
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemConfigUpdateRequest"
            examples:
              updateSearchProvider:
                value:
                  updatedBy:
                    userId: user_demo
                    displayName: Demo Operator
                  values:
                    SEARCH_TOOL_TYPE: BochaAPI
                    BOCHA_WEB_SEARCH_API_KEY: sk-new-value
      responses:
        "200":
          description: Updated masked config.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigResponse"
        default:
          $ref: "#/components/responses/Error"
  /logs:
    get:
      tags: [System]
      summary: Read runtime logs.
      operationId: listLogs
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: source
          in: query
          schema:
            $ref: "#/components/schemas/LogSource"
        - name: level
          in: query
          schema:
            $ref: "#/components/schemas/LogLevel"
        - name: tail
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 2000
            default: 300
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Log lines.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsResponse"
        default:
          $ref: "#/components/responses/Error"
  /search:
    post:
      tags: [Search]
      summary: Dispatch a query to selected engines.
      operationId: createSearchRun
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRunRequest"
      responses:
        "202":
          description: Search run accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchRunResponse"
        default:
          $ref: "#/components/responses/Error"
  /report-templates:
    get:
      tags: [Reports]
      summary: List report templates.
      operationId: listReportTemplates
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Report templates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTemplatesResponse"
        default:
          $ref: "#/components/responses/Error"
  /report-tasks:
    get:
      tags: [Reports]
      summary: List report tasks.
      operationId: listReportTasks
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskStatusFilter"
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Report task page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTaskListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Reports]
      summary: Create a report generation task.
      operationId: createReportTask
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportTaskRequest"
            examples:
              policyTopic:
                value:
                  topic: 养老服务发展趋势
                  templateId: daily-monitoring
                  sourceScope:
                    searchRunId: search_20260522_001
                    includeForumLog: true
                  outputFormats: [html, md, pdf]
                  owner:
                    userId: user_demo
                    displayName: Demo Operator
      responses:
        "202":
          description: Task accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTaskResponse"
        default:
          $ref: "#/components/responses/Error"
  /report-tasks/{taskId}:
    get:
      tags: [Reports]
      summary: Get a report task.
      operationId: getReportTask
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Report task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTaskResponse"
        default:
          $ref: "#/components/responses/Error"
  /report-tasks/{taskId}/events:
    get:
      tags: [Reports]
      summary: Stream report task progress with Server-Sent Events.
      operationId: streamReportTaskEvents
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskId"
        - name: Last-Event-ID
          in: header
          schema:
            type: string
          description: Optional SSE resume cursor.
      responses:
        "200":
          description: SSE stream of ReportTaskEvent messages.
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                progress:
                  summary: Progress event
                  value: |
                    id: 17
                    event: progress
                    data: {"id":"17","type":"progress","taskId":"report_001","timestamp":"2026-05-22T10:10:00Z","payload":{"status":"running","progress":52,"stage":"agent_running","message":"Generating chapter 3"}}
  /report-tasks/{taskId}:cancel:
    post:
      tags: [Reports]
      summary: Cancel a running report task.
      operationId: cancelReportTask
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskId"
      responses:
        "202":
          description: Cancel accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTaskResponse"
        default:
          $ref: "#/components/responses/Error"
  /report-tasks/{taskId}/result:
    get:
      tags: [Reports]
      summary: Get report result metadata and preview URL.
      operationId: getReportResult
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Report result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResultResponse"
        default:
          $ref: "#/components/responses/Error"
  /report-tasks/{taskId}/exports/{format}:
    get:
      tags: [Reports]
      summary: Download or create an exported report artifact.
      operationId: exportReport
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskId"
        - name: format
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ReportFormat"
        - name: optimizeLayout
          in: query
          schema:
            type: boolean
            default: true
          description: Applies to PDF exports.
      responses:
        "200":
          description: Export file stream.
          content:
            text/html:
              schema:
                type: string
            text/markdown:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary
        "202":
          description: Export is queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportExportJobResponse"
        default:
          $ref: "#/components/responses/Error"
  /crawler-strategies:
    get:
      tags: [Crawlers]
      summary: List reusable crawler strategies.
      operationId: listCrawlerStrategies
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Strategy list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerStrategyListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Crawlers]
      summary: Create a crawler strategy.
      operationId: createCrawlerStrategy
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrawlerStrategyInput"
            examples:
              wbStrategy:
                summary: Preserve a platform-bound Weibo policy.
                value:
                  name: 微博每日采集
                  runMode: deep_sentiment
                  platformPolicies:
                    - platformId: wb
                      enabled: true
                      crawlDepth: 3
                      maxKeywords: 10
                      maxNotesPerKeyword: 20
                      maxCommentsPerNote: 50
                      keywords: [养老服务, 医保]
                      keywordSource: manual
                      frequency:
                        mode: daily
                        timezone: Asia/Shanghai
                      loginType: qrcode
                      headless: true
      responses:
        "201":
          description: Strategy created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerStrategyResponse"
        default:
          $ref: "#/components/responses/Error"
  /crawler-tasks:
    get:
      tags: [Crawlers]
      summary: List crawler tasks.
      operationId: listCrawlerTasks
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskStatusFilter"
        - name: platform
          in: query
          schema:
            $ref: "#/components/schemas/PlatformId"
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Crawler task page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerTaskListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Crawlers]
      summary: Create a crawler task.
      operationId: createCrawlerTask
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCrawlerTaskRequest"
            examples:
              dailyCrawl:
                value:
                  strategyId: strategy_daily_hot_topics
                  runMode: full_workflow
                  targetDate: "2026-05-22"
                  platforms: [wb, xhs, zhihu]
                  owner:
                    userId: user_demo
                    displayName: Demo Operator
      responses:
        "202":
          description: Task accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerTaskResponse"
        default:
          $ref: "#/components/responses/Error"
  /crawler-tasks/{taskId}:
    get:
      tags: [Crawlers]
      summary: Get a crawler task.
      operationId: getCrawlerTask
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Crawler task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerTaskResponse"
        default:
          $ref: "#/components/responses/Error"
  /crawler-tasks/{taskId}:stop:
    post:
      tags: [Crawlers]
      summary: Stop a crawler task.
      operationId: stopCrawlerTask
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskId"
      responses:
        "202":
          description: Stop accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerTaskResponse"
        default:
          $ref: "#/components/responses/Error"
  /crawler-tasks/{taskId}:retry:
    post:
      tags: [Crawlers]
      summary: Retry a failed or stopped crawler task.
      operationId: retryCrawlerTask
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TaskId"
      responses:
        "202":
          description: Retry accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerTaskResponse"
        default:
          $ref: "#/components/responses/Error"
  /platforms:
    get:
      tags: [Platforms]
      summary: List supported crawling platforms.
      operationId: listPlatforms
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Platform list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformListResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /platforms/{platformId}/identity-lists/{ruleId}:
    delete:
      tags: [Platforms]
      summary: Delete an identity allow/block rule.
      operationId: deletePlatformIdentityRule
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/PlatformId"
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Rule deleted.
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: X-Workspace-Id
      in: header
      required: true
      description: Tenant/workspace scope for all SaaS resources.
      schema:
        type: string
        example: workspace_demo
...
    ComponentId:
      name: componentId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ComponentId"
...
  schemas:
...
    SystemComponentsResponse:
      type: object
      required: [success, components]
      properties:
        success:
          type: boolean
          const: true
        components:
          type: array
          items:
            $ref: "#/components/schemas/SystemComponent"
    ComponentActionResponse:
      type: object
      required: [success, component]
      properties:
        success:
          type: boolean
          const: true
        component:
          $ref: "#/components/schemas/SystemComponent"
        message:
          type: string
...